+++ src/main/java/com/sdl/selenium/web/utils/WebLocatorConfigReader.java
@@ -15,7 +15,7 @@ public class WebLocatorConfigReader extends PropertiesReader {
             "\n #accepted values for searchType: " + Arrays.asList(SearchType.values()) +
             "\n weblocator.defaults.searchType=CONTAINS" +
             "\n weblocator.defaults.labelPosition=//following-sibling::*//" +
-            "\n weblocator.defaults.value.length=-1" +
+            "\n weblocator.min.chars.toType=-1" +
             "\n driver.autoClose=true" +
             "\n driver.implicitlyWait=100";